fix(jira): extract jira-form-state to satisfy eslint file/complexity limits
deriveFormState pushed jira-settings.tsx over the 600-line file limit and had
cyclomatic complexity 18 (>15). Move FormState/configToForm/savedSecretMatches/
deriveFormState into jira-form-state.ts and split the gating/validation into
jiraDisableSave + jiraInvalidReason helpers.
